requirements.txt
Django==1.5.4
boto==2.3.0
django-jsonfield==0.9.10
django_extensions==1.2.5

